Fixed delete issue; Hide menu icon; Added helper class;
Fixed delete issue: the issue was, when youre in a location & you delete an item, then you go to press a btn to display the same shelf the deleted item was. The item reappears again. It is because the List didnt get updated!
Hide add menu icon: this feature will hide the add button when the location is "sold or missing". 
Helper class: this will be used to add items to the database so I wouldnt have to keep on writing exact same code.